refactor: extract helper funcs
