+++ src/move-ts-indexer/util/shared-interfaces.ts
@@ -1,0 +1,5 @@
+export interface Reference {
+  path: string;
+  specifiers: string[];
+  isExport?: boolean;
+}
